requirements.txt
openai==1.12.0
python-dotenv==1.0.0
pypdf==4.0.1
gradio==4.44.1
pydantic==2.5.3
scikit-learn==1.4.1.post1
tenacity==8.2.3
groq==0.9.0
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
huggingface_hub==0.24.6